rename scrape_config.py to requests.py
