+++ sites/washington_city_paper.py
@@ -7,7 +7,7 @@ import pandas as pd
 from bs4 import BeautifulSoup
 from requests.models import Response
 
-from job.helpers.scrape_config import ScrapeConfig
+from job.helpers.requests import ScrapeConfig
 from job.strategies.collaborative_filtering import CollaborativeFiltering, TrainParamsCF
 from job.strategies.popularity import Popularity
 from lib.events import Event
